backend/requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Supabase client
supabase==2.8.0

# Stripe
stripe==11.0.0

# Monitoring (optional - app works without it)
# sentry-sdk[fastapi]==2.19.0

# HTTP client for external requests
httpx==0.27.2

# Environment and security
python-dotenv==1.0.1
PyJWT==2.9.0